selftest/powerpc/pmu/: Add interface test for mmcr0 exception bits
The testcase uses "instructions" event to verify two
bits(PMAE and PMAO) in Monitor Mode Control Register 0
(MMCR0). At the time of interrupt, pmae bit ( which enables
performance monitor exception ) is expected to be cleared
and pmao (which indicates performance monitor alert) bit
is expected to be set in MMCR0. And testcases handles
these checks.

Signed-off-by: Athira Rajeev <atrajeev@linux.vnet.ibm.com>